DOC-216. Refactor modules
